fixed problem with 13 segments
When the number of segments on screen was 13 - the application filled all segments with one color except one segment. The bug is partially removed by changing rotation of an arc by angle to setting boundaries for each particular arc in degrees.
The applicatoin still have such problem for 81,85,162,170.175 ... segments which yet impractical to fix.